requirements.txt
gTTS
pydub
lyricsgenius
discord.py
giphy_client
praw
chuck-norris-python
youtube-dl
wikipedia